Revert "Apparently we can't support size_t's z modifier in formats."
This reverts (most of) commit ef3b5fe24263f47db8ef937f450cd3655c1b59b4.

BIG FAT NOTE: We provide a C99 snprintf(). So our Log() function and
              most other libutils facilies *are perfectly good* with C99
              modifiers. You should only avoid C99 modifiers when
              calling non-snprintf functions directly from libc, e.g. an
              fprintf() call.

Conflicts:
	cf-serverd/tls_server.c
	libpromises/dbm_tokyocab.c
	libpromises/eval_context.c
	libpromises/evalfunction.c
	libpromises/policy.c
	libpromises/promises.c

+++ libcfnet/communication.c
@@ -146,9 +146,9 @@ int Hostname2IPString(char *dst, const char *hostname, size_t dst_size)
 
     if (dst_size < CF_MAX_IP_LEN)
     {
-        ProgrammingError("Hostname2IPString got %llu, needs at least"
+        ProgrammingError("Hostname2IPString got %zu, needs at least"
                          " %d length buffer for IPv6 portability!",
-                         (unsigned long long)dst_size, CF_MAX_IP_LEN);
+                         dst_size, CF_MAX_IP_LEN);
     }
 
     ret = getaddrinfo(hostname, NULL, &query, &response);
